requirements.txt
fastapi
uvicorn
gunicorn # for run in prod # gunicorn main:app -w 4 -k uvicorn.workers.UvicornWorker
sqlalchemy
aiosqlite
sqlmodel
passlib
pytz
python-jose[cryptography]
email_validator
python-multipart